feat: add Ruby 3.3 Dockerfile
This commit adds a new Dockerfile for Ruby 3.3-alpine. It sets the environment variable `CODECRAFTERS_DEPENDENCY_FILE_PATHS` to "Gemfile,Gemfile.lock" and sets the working directory to `/app`. The `Gemfile` and `Gemfile.lock` are copied to the working directory, and then the `bundle install` command is run to install the dependencies.